refactor: extract hero components
