--- apps/watch/pages/videos/index.tsx
@@ -1,8 +1,0 @@
-import { ReactElement } from 'react'
-import { Videos } from '../../src/components/Videos/Videos'
-
-function VideoPage(): ReactElement {
-  return <Videos />
-}
-
-export default VideoPage
