+++ apps/watch/src/components/Hero/index.ts
@@ -1,0 +1,3 @@
+export { HomeHero } from './HomeHero/HomeHero'
+export { SimpleHero } from './SimpleHero/SimpleHero'
+export { VideoHero } from './VideoHero/VideoHero'
